Refactor YouTube OAuth continuation to inline implementation in page
